Code clean up.
